composer/pyproject.toml
[project]
name = "composer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-cncf-kubernetes==10.0.1",
    "apache-airflow[cncf-kubernetes,google]==2.10.2",
    "pytest>=8.3.4",
]
